moving load_dataset out of scripts
