+++ tgscripts/tf_od_api_eval/evaluate_fo.py
@@ -1,13 +1,9 @@
 """
 
 """
-import pandas as pd
-
 from fiftyone.utils.tfodeval import evaluate_dataset
 
 import fiftyone as fo
-
-pd.set_option("display.max_columns", 500)
 
 ###############################################################################
 
@@ -18,18 +14,6 @@ CLASS_LABELMAP = "/Users/tylerganter/data/open-images-dataset/object_detection/d
 dataset = fo.load_dataset("open-images-v4-test")
 
 evaluate_dataset(dataset, CLASS_LABELMAP, predictions_field_name="faster_rcnn")
-
-###############################################################################
-
-import sys
-
-sys.exit()
-
-import fiftyone as fo
-
-dataset = fo.load_dataset("open-images-v4-test")
-
-s = fo.launch_app(dataset=dataset)
 
 for sample in dataset:
     sample["faster_rcnn_TP"] = fo.Detections(
@@ -55,3 +39,15 @@ for sample in dataset:
     )
 
     sample.save()
+
+###############################################################################
+
+import sys
+
+sys.exit()
+
+import fiftyone as fo
+
+dataset = fo.load_dataset("open-images-v4-test")
+
+s = fo.launch_app(dataset=dataset)
